mv surface and root_surface to ops_batoms
