+++ batoms/__init__.py
@@ -95,8 +95,6 @@ classes_ops = [
     add_object.AddMolecule,
     add_object.AddBulk,
     add_object.AddAtoms,
-    add_object.AddSurface,
-    add_object.AddRootSurface,
     add_surface.BuildSurfaceFCC100,
     add_surface.BuildSurfaceFCC110,
     add_surface.BuildSurfaceFCC111,
@@ -123,6 +121,8 @@ classes_ops = [
     ops_batoms.ApplyCell,
     ops_batoms.ApplyTransform,
     ops_batoms.ApplyBoundary,
+    ops_batoms.AddSurface,
+    ops_batoms.AddRootSurface,
     ops_io.IMPORT_OT_batoms,
     ops_io.EXPORT_OT_batoms,
     gui_batoms.Batoms_PT_prepare,
