+++ batoms/ops/ops_batoms.py
@@ -5,6 +5,7 @@ from bpy.props import (
     BoolProperty,
     FloatVectorProperty,
     FloatProperty,
+    IntProperty,
     IntVectorProperty
 )
 from batoms import Batoms
@@ -47,7 +48,7 @@ class BatomModify(OperatorBatomsEdit):
                          )
     bond: BoolProperty(name="Bond", default=True,
                        )
-    
+
     show: BoolProperty(name="Show", default=True,
                        )
 
@@ -153,4 +154,95 @@ class ApplyBoundary(OperatorBatoms):
         batoms = Batoms(label=context.object.batoms.label)
         batoms.boundary = boundary
         batoms.obj.select_set(True)
+        return {'FINISHED'}
+
+
+class AddSurface(OperatorBatoms):
+    bl_idname = "batoms.surface_add"
+    bl_label = "Add surface"
+    bl_options = {'REGISTER', 'UNDO'}
+    bl_description = ("Add surface")
+
+    label: StringProperty(
+        name="Label", default='surf',
+        description="Label")
+
+    indices: IntVectorProperty(
+        name="Miller indices", size=3, default=(1, 1, 1),
+        soft_min=-10, soft_max=10,
+        description="Miller indices for the plane")
+
+    size: IntVectorProperty(
+        name="Size", size=3, default=(1, 1, 1),
+        min=1, soft_max=10,
+        description="System size in units of the minimal unit cell.")
+
+    vacuum: FloatProperty(
+        name="vacuum", default=5.0,
+        min=0, soft_max=15,
+        description="vacuum")
+
+    nlayer: IntProperty(name="layers", min=1, soft_max=10, default=4)
+
+    termination: StringProperty(
+        name="termination", default='',
+        description="termination")
+
+    def execute(self, context):
+        from ase.build import surface
+        from ase.build.surfaces_with_termination import surfaces_with_termination
+        batoms = Batoms(label=context.object.batoms.label)
+        bulk = batoms.as_ase()
+        if len(self.termination) == 0:
+            atoms = surface(bulk, self.indices,
+                            self.nlayer, self.vacuum)
+        else:
+            atoms = surfaces_with_termination(bulk, self.indices,
+                                              self.nlayer, self.vacuum,
+                                              termination=self.termination)
+        batoms.hide = True
+        label = batoms.label + ''.join(str(i) for i in self.indices)
+        if self.label in bpy.data.collections:
+            self.label = "%s.001" % self.label
+        batoms = Batoms(label=label, from_ase=atoms, movie=True)
+        batoms = batoms*self.size
+        batoms.translate([2, 2, 2])
+        batoms.obj.select_set(True)
+        bpy.context.view_layer.objects.active = batoms.obj
+        return {'FINISHED'}
+
+
+class AddRootSurface(OperatorBatoms):
+    bl_idname = "batoms.root_surface_add"
+    bl_label = "Add root surface"
+    bl_options = {'REGISTER', 'UNDO'}
+    bl_description = ("Add root surface")
+
+    label: StringProperty(
+        name="Label", default='rootsurf',
+        description="Label")
+
+    root: IntProperty(name="root", min=1, soft_max=49, default=3)
+
+    size: IntVectorProperty(
+        name="Size", size=3, default=(1, 1, 1),
+        min=1, soft_max=10,
+        description="System size in units of the minimal unit cell.")
+    
+    eps: FloatProperty(name="eps", min=1e-8, soft_max=1e-3, default=1e-6)
+
+    def execute(self, context):
+        from ase.build import root_surface
+        batoms = Batoms(label=context.object.batoms.label)
+        surf = batoms.as_ase()
+        atoms = root_surface(surf, self.root, eps = self.eps)
+        batoms.hide = True
+        label = batoms.label + '_root'
+        atoms = atoms*self.size
+        if self.label in bpy.data.collections:
+            self.label = "%s.001" % self.label
+        batoms = Batoms(label=label, from_ase=atoms, movie=True)
+        batoms.translate([2, 2, 2])
+        batoms.obj.select_set(True)
+        bpy.context.view_layer.objects.active = batoms.obj
         return {'FINISHED'}
